refactor: split monitor platform implementations
